requirements.txt
exifread
python-ulid
piexif
ffmpeg-python
tqdm